further updates

+++ src/pvn_fmt.c
@@ -48,7 +48,7 @@ size_t pvn_atoz(const char *const s)
 
 char *pvn_ztoa(char *const s, const size_t z)
 {
-  return (s ? ((sprintf((char*)memset(s, 0, (size_t)21u), "%20llu", z) > 0) ? s : (char*)NULL) : s);
+  return (s ? ((sprintf((char*)memset(s, 0, (size_t)21u), "%20llu", (unsigned long long)z) > 0) ? s : (char*)NULL) : s);
 }
 
 float pvn_atos(const char *const s)
